feat #205 : ProfileViewModel 구현
